usb: Add interface specifiers to USB/USB gadget services
This ensures the service can be started as soon as
requested instead of waiting for it to spin on its
own when the "hal" class is started.

This may cause issues specially in the gadget HAL if
the Android framework isn't able to get hold of the
HAL early in boot ( since it may have not started )
and would fail the early configuration.
Thus on the first configuration the HAL would have
stale/improper data and would trigger a USB pull
killing the USB connection, and ADB with it.

Change-Id: I5de309240a7d2f665688b91eb11b36aaea98927e
Signed-off-by: Luca Stefani <luca.stefani.ge1@gmail.com>
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>